HW6/requirements.txt
pytest==8.3.5
python-dotenv==1.1.0
requests==2.32.3
prospector==1.17.1
psycopg2-binary==2.9.10
psycopg2==2.9.10